Updating to positions and targeting
Removed the randomized top elements that were used for telling the sprite apart and implemented more static ones. Targeting was also changed to allow clicking on sprites as well as names. It also now checks to see if the target is dead or not so that you could not continue targeting dead characters.